requirements.txt
# Audio
pydub
ffmpeg-python
mutagen
pillow

# Backend
fastapi
uvicorn[standard]
python-multipart
aiofiles
requests
PyJWT
cryptography

# Frontend
streamlit>=1.42.0
Authlib>=1.3.2

# YouTube upload
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

# Notebooks / experiments (optional)
ipykernel
matplotlib
librosa
numba>=0.51.0,<0.63.0  # Pin numba to avoid llvmlite 0.46.0 (macOS x86_64)
llvmlite==0.45.1        # Last known macOS x86_64 wheel for cp311
numpy
scipy

email-validator
pydantic[email]